lib: add Basic Authentication to HTTP extension
Previously, the only way to enable Basic Authentication was to manually
construct a request object, adding the Authorization header appropriately
and then using do_request. This change allows users to specify a Basic
Authentication user/pass for all request using the head, get and post
functions. Requests returned by request, get_request and post_request
are not affected by the state of the Basic Authentication parameter as
they can be easily altered using the wither pattern shown in the request
documentation example.

No validation of the state of the auth parameter is made and this is the
responsibility of the caller of lib.HTTP/lib.HTTPWithContext.