lib: fix JSON serialisation for non-flat objects (#31)
When an object is converted to native using ConvertToNative with a
native type of map[string]any the resulting native type may be a
concrete map[string]ref.Val which json.Marshal cannot handle. This can
arise when the object is a nested map[string]map[string]any. If it does
occur, Marshal will return a *json.UnsupportedTypeError and we can retry
with the more expensive successive marshalling via the protobuf types.