mito: add limited support for want_more
When used in the Filebeat CEL input, the CEL evaluation can be repeated
with the result of the current evaluation. This was originally added as
a Turing escape hatch for limited use, but it has become a common idiom
for expressing pagination in multistep API requests. The absence of the
want_more option in mito means that the tool is less useful that it
could be when developing CEL programs that use this idiom. So add a
limited version of want_more support.

In the Filebeat input there are a number of checks to ensure that
want_more will not result in runaway computation. This is necessary in
that context as the program is running unattended. In the case of mito,
the program is being used as a validation and debugging tool, and so
will have a human watching its output. So we do not check for too many
loops, or absent events (mito has no concept of events).

The only check is that there is a true state.want_more field in the
final result.