lib: add base64 decode functions
The Go type system already allows specification of []byte values to be
passed into a CEL program, however in the context of free-form
configurations where the types of fields are dynamically typed, it is
not possible to specify that a string of characters that is
syntactically base64 is actually base64. This is the case in filebeat
where the mito libs are used. So add a mechanism to convert a string
type that is syntactically base64 into a bytes.