lib: relax type hints on flatten function
The documented type signature of the flatten function is <list<dyn>> ->
<list<dyn>>, but the implementation was <list<typeV>> -> <list<typeV>>.
This has an impact on uses of flatten where the input is a nested set of
lists, the exact intended use-case. In this situation, the type checker
is told that the returned value is a list of lists..., resulting in
rejection of valid programs. There is a work-around that can be used,
but instead, just make the implementation match the documentation and
intention.