lib: add capacity to limit response body sizes
This is included as some services return responses that are presumably
expected to be consumed as streaming data. This is not possible in the
batch model that CEL uses.

io.CopyN is not used since that does not give an indication that the src
has been truncated without extra work, and the implementation of
io.CopyN is essentially as what is here, except for the mechanism used
to limit the src.