mito,lib,internal/rc: allow configuration of token authentication
Unfortunately we are constrained by compatibility and so cannot return
an error in the case that both Basic Authentication and token
authentication are both configured, so advise against it and populate
both on the basis that it may work. Leave validation of the config to
the client to avoid ever finding out.