mito,mito/lib: implement immediate rate limit application
Previously, the application of rate limits was left entirely to the host
code after the CEL program evaluation has completed. This leaves the
possibility for a collection of work items to deplete the rate
limiter's token bucket in the middle of a CEL program evaluation. This
change adds the possibility of adding a hook for application of rate
limit values on completion of a call to rate_limit, allowing more fine
rate limits to be respected more dynamically.